Adapt to new CMake

+++ ImageReconstruction/Reconstruction/JPetRecoImageTools/JPetFilterInterface.h
@@ -23,7 +23,7 @@ class JPetFilterInterface
 public:
   JPetFilterInterface() {}
   virtual ~JPetFilterInterface(){};
-  virtual double operator()(int) = 0;
+  virtual double operator()(double) = 0;
 };
 
 #endif /*  !_JPetFilterInterface_H_ */
